+++ src/main/java/com/Ayush/Cryptik/service/CoinGeckoService.java
@@ -1,11 +1,10 @@
 package com.Ayush.Cryptik.service;
 
 import java.util.*;
-import java.math.BigDecimal;
 
 import reactor.core.publisher.Mono;
 
 public interface CoinGeckoService {
-    Mono<Map<String, BigDecimal>> getRealTimePrices(List<String> symbols);
+    Mono<Map<String, Integer>> getRealTimePrices(List<String> symbols);
     Mono<Map<String, Object>> getHistoricalData(String symbol, String date);
 }
